+++ src/Http/Controllers/Controller.php
@@ -5,13 +5,14 @@ namespace Laymont\Shicontstand\Http\Controllers;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Foundation\Bus\DispatchesJobs;
 use Illuminate\Foundation\Validation\ValidatesRequests;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Routing\Controller as BaseController;
 
 class Controller extends BaseController
 {
     use AuthorizesRequests, DispatchesJobs, ValidatesRequests;
 
-    public function responseProhibited(): \Illuminate\Http\JsonResponse
+    public function responseProhibited(): JsonResponse
     {
         return response()->json([
             'message' => 'Prohibited Resource',