+++ src/Http/Resources/SizeCodeResource.php
@@ -3,6 +3,7 @@
 namespace Laymont\Shicontstand\Http\Resources;
 
 use Illuminate\Contracts\Support\Arrayable;
+use Illuminate\Http\Resources\Json\JsonResource;
 use JsonSerializable;
 
 /**
@@ -10,7 +11,7 @@ use JsonSerializable;
  * @property mixed $container_height
  * @property mixed $width
  */
-class SizeCodeResource extends \Illuminate\Http\Resources\Json\JsonResource
+class SizeCodeResource extends JsonResource
 {
     public function toArray($request): array|JsonSerializable|Arrayable
     {